+++ askbot/management/commands/delete_unused_tags.py
@@ -14,8 +14,18 @@ class Command(NoArgsCommand):
         deleted_tags = list()
         for tag in ProgressBar(tags, total, message):
             if not tag.threads.exists():
-                deleted_tags.append(tag.name)
-                tag.delete()
+                #if any user subscribed for the tag and
+                #the user is not blocked, skip deleting the tag
+                marks = tag.user_selections.all()
+                do_delete = True
+                for mark in marks:
+                    if not mark.user.is_blocked():
+                        do_delete = False
+                        break
+
+                if do_delete:
+                    deleted_tags.append(tag.name)
+                    tag.delete()
 
         if deleted_tags:
             found_count = len(deleted_tags)