+++ askbot/models/question.py
@@ -237,15 +237,16 @@ class ThreadManager(BaseQuerySetManager):
 
         #todo: this is handled in signal because models for posts
         #are too spread out
-        signals.post_updated.send(
-            post=question,
-            updated_by=author,
-            newly_mentioned_users=parse_results['newly_mentioned_users'],
-            timestamp=added_at,
-            created=True,
-            diff=parse_results['diff'],
-            sender=question.__class__
-        )
+        if revision.revision > 0:
+            signals.post_updated.send(
+                post=question,
+                updated_by=author,
+                newly_mentioned_users=parse_results['newly_mentioned_users'],
+                timestamp=added_at,
+                created=True,
+                diff=parse_results['diff'],
+                sender=question.__class__
+            )
 
         return thread
 
@@ -818,11 +819,32 @@ class Thread(models.Model):
             return answers[0].id
         return None
 
-    def get_latest_post(self):
-        """returns latest non-deleted post"""
-        if askbot_settings.GROUPS_ENABLED:
-            raise NotImplementedError()
-        return self.posts.filter(deleted=False).order_by('-added_at')[0]
+    def get_latest_revision(self, user=None):
+        #todo: add denormalized field to Thread model
+        from askbot.models import Post, PostRevision
+        posts_filter = {
+            'thread': self,
+            'post_type__in': ('question', 'answer'),
+            'deleted': False
+        }
+
+        if user and askbot_settings.GROUPS_ENABLED:
+            #get post with groups shared with having at least 
+            #one of the user groups
+            #of those posts return the latest revision
+            posts_filter['groups__in'] = user.get_groups()
+
+        posts = Post.objects.filter(**posts_filter)
+        post_ids = list(posts.values_list('id', flat=True))
+
+        revs = PostRevision.objects.filter(
+                                post__id__in=post_ids,
+                                revision__gt=0
+                            )
+        try:
+            return revs.order_by('-id')[0]
+        except IndexError:
+            return None
 
     def get_primary_site_ids(self):
         """takes threads spaces and returns site ids
@@ -1202,6 +1224,7 @@ class Thread(models.Model):
                                                 post__id__in=post_ids,
                                                 revision=0
                                             )
+
             #get ids of posts that we need to patch with suggested data
             if len(suggested_revs):
                 #find posts that we need to patch
@@ -1218,7 +1241,7 @@ class Thread(models.Model):
                             found.update(find_posts(comments, need_ids))
                     return found
 
-                suggested_post_ids = set([rev.post_id for rev in suggested_revs])
+                suggested_post_ids = [rev.post_id for rev in suggested_revs]
 
                 question = post_data[0]
                 answers = post_data[1]
@@ -1237,7 +1260,10 @@ class Thread(models.Model):
                     rev = rev_map[post_id]
                     #patching work
                     post.text = rev.text
-                    post.html = post.parse_post_text()['html']
+                    parse_data = post.parse_post_text()
+                    post.html = parse_data['html']
+                    post.summary = post.get_snippet()
+
                     post_to_author[post_id] = rev.author_id
                     post.set_runtime_needs_moderation()
 
@@ -1264,6 +1290,7 @@ class Thread(models.Model):
                         rev = rev_map[post.id]
                         post.text = rev.text
                         post.html = post.parse_post_text()['html']
+                        post.summary = post.get_snippet()
                         post_to_author[post.id] = rev.author_id
                         if post.is_comment():
                             parents = find_posts(all_posts, set([post.parent_id]))
@@ -1338,6 +1365,13 @@ class Thread(models.Model):
         post_to_author = dict()
         question_post = None
         for post in thread_posts:
+
+            #precache some revision data
+            first_rev = post.get_earliest_revision()
+            last_rev = post.get_latest_revision()
+            first_rev.post = post
+            last_rev.post = post
+
             #pass through only deleted question posts
             if post.deleted and (not post.is_question()):
                 continue
@@ -1865,7 +1899,7 @@ class Thread(models.Model):
 
         return last_updated_at, last_updated_by
 
-    def get_summary_html(self, search_state=None, visitor = None):
+    def get_summary_html(self, search_state=None, visitor=None):
         html = self.get_cached_summary_html(visitor)
         if not html:
             html = self.update_summary_html(visitor)